requirements.txt
google-genai>=1.0,<2.0
panel>=1.3,<2.0
python-dotenv>=1.0,<2.0
pandas>=2.0,<3.0
tabulator>=1.0,<2.0
pytest>=7.0,<9.0
chardet==4.0.0
